requirements.txt
# Core Framework - Updated for compatibility
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# Document Processing
PyMuPDF==1.23.14
pdfplumber==0.10.3
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
opencv-python==4.8.1.78
poppler-utils

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Vector Database
qdrant-client==1.7.0

# AI/ML
openai==1.10.0
tiktoken==0.5.2

# Data Processing
pandas==2.1.4
numpy==1.26.3
pydantic==2.5.3
pydantic-settings==2.1.0

# Web Framework (FastAPI)
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.4
black==23.12.1
flake8==7.0.0
